+++ api-svc/raid-api/src/main/java/au/org/raid/api/service/raid/RaidService.java
@@ -1,6 +1,7 @@
 package au.org.raid.api.service.raid;
 
 import au.org.raid.api.dto.RaidPermissionsDto;
+import au.org.raid.api.dto.legacy.RaidDtoFactory;
 import au.org.raid.api.exception.InvalidVersionException;
 import au.org.raid.api.exception.ResourceNotFoundException;
 import au.org.raid.api.exception.ServicePointNotFoundException;
@@ -62,6 +63,7 @@ public class RaidService {
     private final ObjectMapper objectMapper;
 
     private final RaidRepository raidRepository;
+    private final RaidDtoFactory raidDtoFactory;
 
     @Transactional
     public RaidDto mint(
@@ -281,5 +283,19 @@ public class RaidService {
                 .orElseThrow(() -> new ServicePointNotFoundException(servicePointId));
 
         dataciteSvc.update(raid, handle.toString(), servicePointRecord.getRepositoryId(), servicePointRecord.getPassword());
+    }
+
+    public List<RaidDto> findAllEmbargoed() {
+        final var raidRecords = raidRepository.findAllEmbargoed();
+
+        final var raids = new ArrayList<RaidDto>();
+
+        for (final var record : raidRecords) {
+            final var raidDto = raidHistoryService.findByHandle(record.getHandle())
+                    .orElseThrow(() -> new ResourceNotFoundException(record.getHandle()));
+            raids.add(raidDto);
+        }
+
+        return raids;
     }
 }